Rename the Waves C++ header
